### 2024-11-15 #7487 @warrickball
@ChunHuangPhy — Thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on.

I also noticed that our system hasn't automatically recognised your MIT License. I'm not quite sure why not but it might be because your specific text differs from the standard version. Specifically, I think these lines
```
Copyright (c)[2024 Chun Huang, João Cartaxo, Tuhin Malik, Wenli Yuan,
Shashwat Sourav, Xuezhi Liu, Tianzhe Zhou, John Groger, Zhiheng Wang, Nicole Osborn, Nathan Whitsett]
```
might need to be
```
Copyright (c) 2024 Chun Huang, João Cartaxo, Tuhin Malik, Wenli Yuan,
Shashwat Sourav, Xuezhi Liu, Tianzhe Zhou, John Groger, Zhiheng Wang, Nicole Osborn, Nathan Whitsett
```
and you might even need to put all the authors back on one line. Have a look at other MIT Licensed code you might know of or browse past JOSS articles to see what might be the issue here, then run `@editorialbot check repository` to rerun our checks.

Also, the paper is failing to build at least because of a YAML error in the metadata, which is also flagged by the GitHub UI if you [view the file there](https://github.com/ChunHuangPhy/CompactObject/blob/b5705dceece84acb4c4cb6f40b6ddfb6100d2a84/JOSS/paper.md). Specifically, I think the `tags` should be indented by two spaces, not four.

If you think you've fixed this, you can try rebuilding the paper with `@editorialbot generate pdf`.
