### 2024-11-16 #7487 @warrickball
I appreciate your efforts here. I'll try to debug why the paper doesn't compile.

In the meantime, does someone in @openjournals/dev know why the pipeline isn't detecting the MIT License? Is it because the authors are spread over a line break? Aside from quotation-mark styles, the text looks exactly correct.
