### 2024-11-16 a5dd9cf
Fix metadata in JOSS paper source
