+++ InferenceWorkflow/__init__.py
@@ -1,2 +1,2 @@
 
-__version__ = "1.9.9"
+__version__ = "2.0.0"
