+++ setup.py
@@ -10,7 +10,7 @@ def get_requires():
 
 setup(
     name="CompactObject-TOV",
-    version="1.9.9",
+    version="2.0.0",
     packages=find_packages(),
     # install_requires=get_requires(),  # Automatically adds requirements
     # other setup parameters
