+++ TOVsolver/__init__.py
@@ -1,4 +1,4 @@
 
-__version__ = "1.9.9"
+__version__ = "2.0.0"
 
 
