### 2024-11-19 #7487 @xuanxu
> Do we need to change the MIT license to make it pass the test?

No need. The bot check is just a tool to help editors, reading the license file I think is clear it's a valid MIT license.
