### 2024-11-25 c9a10e8
add strangeon arxiv
