+++ _sources/starthere.rst.txt
@@ -1,0 +1,127 @@
+Start Here!
+===========
+
+Welcome to **CompactObject**! This guide will help you get started after successfully installing the package and activating your virtual environment. Follow the steps below to achieve your scientific goals using the functionalities of the **CompactObject** package.
+
+Prerequisites
+-------------
+
+- **Package Installation:** Ensure you have installed **CompactObject** successfully, see installation instructions here: `Installation <https://chunhuangphy.github.io/CompactObject/installation.html>`_ .
+- **Virtual Environment:** Activate your virtual environment where **CompactObject** is installed , also see installation instructions here: `Installation <https://chunhuangphy.github.io/CompactObject/installation.html>`_ .
+
+Depending on Your Scientific Goal
+---------------------------------
+
+1. Understanding Basic Concepts
+   ----------------------------
+
+   If you are new to **CompactObject** and wish to familiarize yourself with its basic concepts and introduction:
+
+   - **Index Page:** Explore the `Index Page <https://chunhuangphy.github.io/CompactObject/index.html>`_ for an overview of the package.
+   - **README:** Review the `README <https://github.com/ChunHuangPhy/CompactObject/blob/main/README.md>`_ on GitHub to understand the package’s purpose and features.
+
+2. Generating Equations of State (EOS)
+   ------------------------------------
+
+   To use **CompactObject** for generating equations of state:
+
+   - **EOS Generators Tutorial:** Start with the `EOS Generators Tutorial <https://chunhuangphy.github.io/CompactObject/test_EOSgenerators.html>`_, which showcases all currently integrated EOS computations in the package. The supported EOS models include:
+
+     - **Relativistic Mean Field (RMF) Model**
+     - **Strangeon Star Model**
+     - **Quark Star Model:** MIT Bag Model
+     - **Polytropic Model**
+     - **Speed of Sound Model**
+     - **Density-Dependent EOS Model**
+     - *...and more*
+
+3. Using the Tolman–Oppenheimer–Volkoff (TOV) Solver
+   -------------------------------------------------
+
+   To utilize the TOV solver integrated within **CompactObject**:
+
+   - **TOV Solver Tutorial:** Refer to the `TOV Solver Tutorial <https://chunhuangphy.github.io/CompactObject/test_TOVsolver.html>`_. This tutorial demonstrates how to perform TOV computations using table-based EOS.
+   - **Inputting an Array:** If you prefer to input an array instead of a table, refer back to the `EOS Generators Tutorial <https://chunhuangphy.github.io/CompactObject/test_EOSgenerators.html>`_. The quantities currently available are:
+
+     - **Mass**
+     - **Radius**
+     - **Tidal Deformability**
+
+4. Performing Bayesian Inference on the EOS
+   -----------------------------------------
+
+   **CompactObject** offers robust tools for Bayesian inference related to neutron star EOS. There are two primary scenarios:
+
+   a. **Using Integrated EOS for Inference**
+
+      If you wish to perform Bayesian inference using the EOS models already integrated into **CompactObject**, follow these examples:
+
+      1. **RMF EOS Inference Pipeline:** Explore the `RMF EOS Inference Pipeline <https://chunhuangphy.github.io/CompactObject/test_Inference.html>`_.
+      2. **density-dependent Hartree EOS Inference Pipeline:** Explore the `DDH EOS Inference Pipeline <https://chunhuangphy.github.io/CompactObject/test_Bayesian_inference_DDH.html>`_.
+      3. **Strangeon EOS Inference Pipeline:** Check out the `Strangeon EOS Inference Pipeline <https://chunhuangphy.github.io/CompactObject/test_Bayesian_inference_Strangeon_EOS.html>`_.
[…]
