+++ _sources/RMF_EOS.rst.txt
@@ -1,0 +1,9 @@
+.. _RMF_EOS:
+
+RMF EOS solver 
+=====================
+
+Functions to compute RMF Equation of state from given parameters.
+
+.. automodule:: EOSgenerators.RMF_EOS
+   :members:
