+++ _sources/main.rst.txt
@@ -1,0 +1,12 @@
+.. _main:
+
+Main functions
+=====================
+
+Controls most functionality of package by calling the following functions.
+* :ref:'EoS_importing'
+* :ref: 'TOV_solving'
+* :ref: '__s'
+
+.. automodule:: TOVsolver.main
+   :members:
