+++ _sources/Compose_eos.rst.txt
@@ -1,0 +1,9 @@
+.. _compose_EOS:
+
+Omplement CompOSE dataset 
+=========================
+
+Functions to Implement CompOSE dataset.
+
+.. automodule:: EOSgenerators.Compose_eos
+   :members:
