+++ _sources/RMF_DDH.rst.txt
@@ -1,0 +1,9 @@
+.. _RMF_DDH_EOS:
+
+DDH RMF EOS solver 
+=====================
+
+Functions to compute DDH RMF Equation of state from given parameters.
+
+.. automodule:: EOSgenerators.RMF_DDH
+   :members:
