+++ _sources/index.rst.txt
@@ -1,0 +1,87 @@
+.. module:: CompactObject
+
+.. include:: intro.rst
+
+.. toctree::
+   :maxdepth: 2
+   :caption: User Guide
+   
+   starthere
+   installation
+   publications
+   cluster
+   
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Development
+   
+   Contributor
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials
+
+   test_EOSgenerators
+   test_TOVsolver
+   test_Inference
+   test_Bayesian_inference_DDH
+   test_Bayesian_inference_Strangeon_EOS
+   test_Bayesian_inference_MITbag_EOS
+   test_Bayesian_inference_SpeedOfSound_EOS
+   test_Inference_polytrope
+   DownloadCompose
+   UnitConventionForDeveloper
+
+.. toctree::
+   :maxdepth: 2
+   :caption: MCMC inference workflow
+
+   EMCEE_RMF
+   EMCEE_strangeon
+   EMCEE_MIT_bag
+   EMCEE_polytrope
+
+.. toctree::
+   :maxdepth: 2
+   :caption: EOS generators module
+
+   crust_EOS
+   RMF_EOS
+   RMF_DDH
+   Strangeon_EOS
+   MITbag_EOS
+   Polytrope_EOS
+   SpeedofSound_EOS
+   Compose_eos
+   fastRMF_EoS
+
+.. toctree::
+   :maxdepth: 2
+   :caption: TOV solver module
+
+   EoS_import
+   main
+   TOV_solver
+   speed_of_sound
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Bayesian Inference module
+
+   BayesianSampler
+   prior
+   Likelihood
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Postprocessing module
+
+   Cornerplot
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`

[…]
