### 2025-02-01 #7487 @warrickball
Hi again @ChunHuangPhy, and sorry for the long wait. I'll take this on as an editor and start looking for (at least) two reviewers. If you have anyone you'd like to suggest, let me know here. You may wish to have a look at authors of [previous JOSS papers](https://joss.theoj.org/papers) and in our [reviewer database](https://reviewers.joss.theoj.org/lookup). Note that we do have a [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).

I'll start looking largely by these routes in the meantime.
