### 2025-02-07 #7487 @warrickball
Hi @ChunHuangPhy, I've identified two reviewers and am ready to start the review.

For transparency, note that both reviewers have indicated potential [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) that I've acknowledged and am happy to waive. @kelslund is on a conference organising committee that's vice-chaired by one of the authors and @rhaas80 co-authored a large [review article in 2024](https://ui.adsabs.harvard.edu/abs/2024LRR....27....3K) with another. I don't believe these conflicts will influence their reviews but I'll wait a few days to start the review in case you wish to comment.
