### 2025-02-07 #7759 @editorialbot
[REVIEW]: CompactObject: An open-source Python package for full-scope neutron star equation of state inference
**Submitting author:** @ChunHuangPhy (<a href="http://orcid.org/0000-0001-6406-1003">Chun Huang</a>)
**Repository:** https://github.com/ChunHuangPhy/CompactObject
**Branch with paper.md** (empty if default branch):
**Version:** v1.9.9
