### 2025-02-17 #7759 @warrickball
Hi @kelslund & @rhaas80, I was wondering if you'd had a chance to start your reviews? We recommend creating your checklist with `@editorialbot generate my checklist` and commenting on issues as soon as you come across them, rather than waiting until you've looked at everything for a traditional, monolithic review.

Also, let me know if you're having any issues with the review process itself and I'll try to help resolve them.
