### 2025-02-21 #7759 @warrickball
Hi @rhaas80 — I realise my last message might not have been clear. Each reviewer should generate their own checklist, so you should also comment with `@editorialbot generate my checklist`, even if you haven't started your review.
