### 2025-03-10 #7759 @warrickball
Hi everyone! I'm just checking in on this review. @ChunHuangPhy, note (if you haven't already) that @rhaas80 has [opened several issues](https://github.com/ChunHuangPhy/CompactObject/issues?q=is%3Aissue%20state%3Aopen%20author%3Arhaas80) in the project repo itself, which is a perfectly reasonable way to proceed, and quite useful for separating different elements of the review. Please have a look at those if you haven't already.

@kelslund, is there anything you need my help on? Note that we encourage reviewers to comment on issues and ask questions as soon as they come up. Unlike mainstream journals, we don't expect a single monolithic review, and find it more efficient when the authors can comment on one part of the review while the reviewers move on to the next.
