### 2025-03-14 #7759 @kelslund
Hi, sorry, I have actually been having problems with the some of the notebooks in the Test_case folder. @ChunHuangPhy I proceeded with the installation with seemingly no problems, but then it seems like I am having trouble with some of the libraries.
