### 2025-03-15 #7759 @warrickball
Thanks @kelslund. If there's anything unclear about the tests, that's something @ChunHuanPhy and the authors should work on. Could you be more specific about the problems, either here or in [an issue in the code repo](https://github.com/ChunHuangPhy/CompactObject/issues)?
