### 2025-04-01 ChunHuangPhy/CompactObject#130 @kelslund
test_inference hangs up with suggested parameters
When running test case notebook test_Inference on a laptop, the code hangs up (1.5 hr+) when the suggested parameters (live_point=2000, max_calls=60000) are used in the step where the sampler is set up. Output includes:
```
home/lindsey/.local/lib/python3.8/site-packages/InferenceWorkflow/Likelihood.py:115: RuntimeWarning: divide by zero encountered in log
  likelihood = np.log(fx)
/home/lindsey/.local/lib/python3.8/site-packages/TOVsolver/solver_code.py:80: RuntimeWarning: overflow encountered in scalar multiply
  dpdr = dpdr / (r * (r - 2.0 * m))
/home/lindsey/.local/lib/python3.8/site-packages/TOVsolver/solver_code.py:79: RuntimeWarning: overflow encountered in scalar multiply
  dpdr = -(eps + pres) * (m + 4.0 * pi * r**3.0 * pres)
/home/lindsey/.local/lib/python3.8/site-packages/TOVsolver/solver_code.py:80: RuntimeWarning: invalid value encountered in scalar divide
  dpdr = dpdr / (r * (r - 2.0 * m))
/home/lindsey/.local/lib/python3.8/site-packages/scipy/integrate/_ode.py:1179: UserWarning: dopri5: step size becomes too small
  warnings.warn('{:s}: {:s}'.format(self.__class__.__name__,
/home/lindsey/.local/lib/python3.8/site-packages/TOVsolver/solver_code.py:284: RuntimeWarning: overflow encountered in scalar divide
  return stateTOV[1] * unit.c**2 / unit.G, sy.t

[ultranest] Sampling 2000 live points from prior ...

/home/lindsey/.local/lib/python3.8/site-packages/scipy/integrate/_ode.py:1179: UserWarning: dopri5: larger nsteps is needed
  warnings.warn('{:s}: {:s}'.format(self.__class__.__name__,
capi_return is NULL
Call-back cb_fcn_in___user__routines failed.
capi_return is NULL
Call-back cb_fcn_in___user__routines failed.
```
Changing max_calls to 10000 or 20000 removes these warnings, but still does not complete within 30 minutes,
