### 2025-04-14 ChunHuangPhy/CompactObject#130 @kelslund
This problem has not been unique to the single notebok. @ChunHuangPhy, the text in the notebooks suggest these test cases were run on a laptop. I am also trying to run these on a laptop but the samplers keep hanging up and not giving results. This has been the case for sample_glitchBayesian and test_Bayesian_inference_SpeedofSound_EOS, so far.
