### 2025-05-08 #7759 @warrickball
Hi @ChunHuangPhy, I see @kelslund opened at least [one issue in the code repo](https://github.com/ChunHuangPhy/CompactObject/issues/130). Have you possibly had a chance to address that, along with the various issues raised by @rhaas80?
