### 2025-05-17 ChunHuangPhy/CompactObject#130 @tuhinbits
## Key Improvements in the TOV Solver

1. **Comprehensive Documentation**
   - Extensively documented function with clear descriptions of parameters, return values, and methodology
   - Added detailed physics notes explaining the numerical approach and potential issues
   - Included explanations of the Taylor expansion used to avoid the central singularity

2. **Adaptive Step Sizing Algorithm**
   - Implemented the physically-motivated adaptive step sizing with citation:
     ```python
     delta = 0.23  # optimal delta from https://articles.adsabs.harvard.edu/pdf/1971ApJ...170..299B
     dr = delta / ((1.0 / stateTOV[1]) * dmdr - (1.0 / stateTOV[0]) * dpdr)
     ```
3. **Multiple Robust Stopping Conditions**
   - Added layered stopping criteria to handle various edge cases:
     ```python
     while (sy.successful() and
            stateTOV[0] > Pmin and
            sy.t < max_radius and
            rho_current > 0.01 * center_rho):
     ```
   - Density threshold condition to catch surface boundary more reliably
   - Maximum radius safety parameter to prevent runaway integration
   - Mass consistency check to detect numerical instabilities:
     ```python
     if stateTOV[1] < prev_mass:
         break
     ```

4. **Surface Region Handling**
   - Special step size reduction near the stellar surface:
     ```python
     if stateTOV[0] < Pmin * 10:
         dr = min(dr, 50.0 * unit.cm)
     ```
   - This directly addresses the warnings about step size becoming too small in steep gradient regions

5. **Improved Initial Conditions**
   - Carefully chosen initial radius (1e-18 * unit.cm) to balance numerical stability concerns
   - Taylor expansion for initial pressure to handle the coordinate singularity at r=0

6. **Error Tolerance Configuration**
   - Configured the ODE integrator with appropriate absolute and relative error tolerances:
     ```python
     sy = ode(TOV, None).set_integrator("dopri5", atol=1e-8, rtol=1e-8)
     ```
   - Using a more suitable integrator (dopri5) for handling stiff differential equations

These improvements collectively enhance the numerical stability, accuracy, and robustness of the neutron star structure calculations across a wide range of equations of state.

## Dynasty Sampler for Notebook Users
Added Dynasty sampler with multiprocessing pool support for efficient notebook execution
Key advantage: Utilizes all CPU cores without requiring MPI configuration
Performance: Achieves near-linear scaling with available cores on a single machine
Maintained UltraNest support for MPI-enabled cluster environments
