### 2025-07-24 #7759 @warrickball
Ah, I see. I only looked at the first example issue I linked above and not more widely. I've left a comment in an issue there that it looks like you need to rebuild the live documentation before it's worth @rhaas80 looking again. I'll comment on some of @kelslund's too, where I think the reviewers need a bit more guidance on what's been done.
