### 2025-07-24 ChunHuangPhy/CompactObject#130 @warrickball
Hi @tuhinbits, these comments are extensive but I don't think it's clear enough for @kelslund what they should expect.

Has the code been modified since they opened the issue? Should they try again now? If they do, what result should they expect? Is the very long scalability analysis comparable to the example they're trying to run?
