### 2026-02-14 #7759 @warrickball
Hi everyone! @kelslund and @rhaas80, I left some remarks in the various issues and it would be great if you could revisit them.

In the meantime, @ChunHuangPhy, could you please update everyone on your progress in the main repo (as I requested [above](https://github.com/openjournals/joss-reviews/issues/7759#issuecomment-3726433643))? JOSS's volunteer time is under signficiant pressure at the moment, so if I don't hear anything here in the next four weeks (by 14 March), I'll assume you're no longer interested in pursuing this submission and mark it as withdrawn. I'm willing to wait longer if you could send any message to indicate you're still engaged with this review.
