### 2026-02-14 #7759 @ChunHuangPhy
Hi everyone, thank you @warrickball  for the reminder and for your continued time on this review.

I’m still actively engaged and definitely interested in continuing this JOSS submission.

Progress update on CompactObject:

1. I made a batch of documentation-focused updates (especially around TOVsolver/solver_code.py, docs/source/solver_code.py, and docs/source/RMF_EOS.py) to improve/clarify function signatures and Returns sections.
2. I removed outdated install/pip-related pieces .
3. I updated README citations for recent newly published paper based on this package .
4. I applied a solver robustness fix in November 2025 (solveTOV_tidal stopping condition logic).
5. I pushed additional definition/docstring cleanup in January 2026, plus package-paper related updates.
6. On the review issues: I have responded on multiple threads and finished incorporating requested changes.

I appreciate your patience and would be grateful to keep the review open. I remain committed to completing this process.
