### 2026-03-22 #7759 @warrickball
Thanks @ChunHuangPhy.

@rhaas80 & @kelslund, rather than comment on the individual issues in repo, I was wondering if you could let me know here if you're still available to continue your reviews? I'm about to go on vacation for basically the next four weeks, and if I haven't heard back by the time I return on April 20, I'll start looking for new reviewers.
