### 2026-08-16 #7759 @warrickball
@ChunHuangPhy, please note the issues that @svisak has also raised and helpfully cross-referenced. The first was opened about 3 weeks ago now.
